Using Adapter Design Pattern in Template Engine
