+++ src/Services/TemplateEngineAdapter.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace ResumeCraft\Services;
+
+interface TemplateEngineAdapter
+{
+    /**
+     * @param array $data
+     * @return string
+     */
+    public function renderTemplate(array $data): string;
+}
